+++ src/component/field/Field.tsx
@@ -4,11 +4,12 @@ import React from 'react';
 
 const textFieldStyle = { width: "155px", marginBottom: "10px", marginRight: "10px", marginLeft: "10px" };
 
-const NeoFieldSelection = ({ label,
+const NeoField = ({ label,
     value,
     style = textFieldStyle,
     choices = [<div key={0}></div>],
     onChange,
+    onClick = (val)=>{},
     numeric = false,
     select = false,
     disabled = undefined,
@@ -30,6 +31,7 @@ const NeoFieldSelection = ({ label,
             disabled={disabled}
             value={value != null ? value : defaultValue}
             // defaultValue={defaultValue}
+            onClick={(e) => {onClick(e.target.textContent)}}
             onChange={(event) => {
                 if (!numeric) {
                     onChange(event.target.value);
@@ -47,4 +49,4 @@ const NeoFieldSelection = ({ label,
 };
 
 
-export default NeoFieldSelection;
+export default NeoField;